rework builder annotations and req for signing key
The builder annotation change is motivated by not wanting to keep up to
date with how to parse the builder metadata.

The signing key annotation change is to make it possible to use cosign
secrets for signing or attesting only without opting into both.

Also add ED25519 since the golang crypto/x509 lib supports all 3 in the
PKCS8 implementation

Signed-off-by: Bohan Chen <bohanc@vmware.com>