split into app and builder image attestations
There's actually 2 images that are completely generated by kpack (i.e.
not brought in by user). The builder image is stiched together from
existing layers, and the app image that is build via the builder image.

While the app image attestation is likely the most important one, we
should still attest that our builder images are not tempered with

Signed-off-by: Bohan Chen <bohanc@vmware.com>